Split up trap sender function
